feat(list): break card components into into list
Break out allows Card component to self state manage.
