+++ src/componant/List/index.css
@@ -1,0 +1,4 @@
+.container {
+  width: 100%;
+  max-width: 300px;
+}
