+++ src/componant/card/index.css
@@ -1,9 +1,8 @@
 .card {
   background: var(--color-white);
-  width: 100%;
-  max-width: 300px;
   border-radius: 10px;
   box-shadow: 0 0 100px 0 rgba(0, 0, 0, 0.25);
+  margin-bottom: 100px;
 }
 
 /* body */
