+++ src/layout/main/index.css
@@ -1,9 +1,13 @@
 .main {
-  width: 100vw;
-  height: 100vh;
+  min-height: 100vh;
   background: var(--color-tomato);
   display: flex;
   flex-direction: column;
   align-items: center;
   padding-top: 100px;
+}
+
+.container {
+  width: 100%;
+  max-width: 300px;
 }
